+++ src/rpc/misc.cpp
@@ -16,6 +16,7 @@
 #include <outputtype.h>
 #include <rpc/blockchain.h>
 #include <rpc/server.h>
+#include <rpc/server_util.h>
 #include <rpc/util.h>
 #include <scheduler.h>
 #include <script/descriptor.h>
